helperFunctions/config_deprecated: Remove module
