--- docsrc/modules/helperFunctions.config_deprecated.rst
@@ -1,7 +1,0 @@
-helperFunctions.config module
-=============================
-
-.. automodule:: helperFunctions.config_deprecated
-   :members:
-   :undoc-members:
-   :show-inheritance:
